Refactor the cache to be more generic.
This actually encompasses a pretty significant reworking of the data model.
Block and inode quotas now have separate entries in the cache which means the
code can be more generic.  This was also an opportunity to separate the
quota information from the account information so now the AccountInfo object
contains a collection of QuotaInfo objects rather than representing that
information directly.  A side effect of this change is that a single account now
has information about all of the filesystems on the system, so if a person is
over quota on multiple filesystems they'll get just a single email that contains
all of that information (rather than getting one email for each filesystem).